refactor: разложил ошибки в отдельные файлы
