+++ .eslintrc
@@ -1,7 +1,6 @@
 {
   "extends": "airbnb-base",
   "rules": {
-    "no-underscore-dangle": ["error",{"allow":["_id"]}],
-    "max-classes-per-file": ["error", 10]
+    "no-underscore-dangle": ["error",{"allow":["_id"]}]
  }
 }
