+++ controllers/movies.js
@@ -1,14 +1,10 @@
 const Movie = require('../models/movie');
-const { classes, names } = require('../errors/index');
+const { names } = require('../errors/index');
 const { StatusCodes } = require('../support/statusCodes');
 const messages = require('../support/messages');
-const { ForbiddenError } = require('../errors/classes');
-
-const {
-  NotFoundError, BadRequestError, /* ForbiddenError, */
-} = classes;
-
-// const defaultPopulation = ['owner'];
+const ForbiddenError = require('../errors/forbiddenerror');
+const NotFoundError = require('../errors/notfounderror');
+const BadRequestError = require('../errors/badrequesterror');
 
 module.exports.createMovie = (req, res, next) => {
   const {
@@ -44,14 +40,12 @@ module.exports.createMovie = (req, res, next) => {
 };
 
 module.exports.getMovies = (req, res, next) => {
-  // Movie.find({})
-  // .populate(defaultPopulation)
   Movie.find({ owner: req.user._id })
     .then((data) => res.send({ data }))
     .catch(next);
 };
 
-// Исправить как-то, не удаляется.
+// удаляется, но остается пустой объект
 module.exports.deleteMovie = (req, res, next) => {
   Movie.findById(req.params.movieId)
     .orFail(() => {
@@ -62,7 +56,7 @@ module.exports.deleteMovie = (req, res, next) => {
         throw new ForbiddenError();
       }
       await movie.remove();
-      return res.send({ message: messages.ok }); // исправлено
+      return res.send({ message: messages.ok });
     })
     .catch((err) => {
       next(err);
