+++ errors/conflicterror.js
@@ -1,0 +1,10 @@
+const { messages } = require('../support/messages');
+
+class ConflictError extends Error {
+  constructor(message = messages.conflict) {
+    super(message);
+    this.statusCode = 409;
+  }
+}
+
+module.exports = ConflictError;
