+++ errors/forbiddenerror.js
@@ -1,0 +1,10 @@
+const { messages } = require('../support/messages');
+
+class ForbiddenError extends Error {
+  constructor(message = messages.forbidden) {
+    super(message);
+    this.statusCode = 403;
+  }
+}
+
+module.exports = ForbiddenError;
