+++ errors/index.js
@@ -1,7 +1,5 @@
-const classes = require('./classes');
 const names = require('./names');
 
 module.exports = {
-  classes,
   names,
 };
