+++ errors/internalservererror.js
@@ -1,0 +1,10 @@
+const { messages } = require('../support/messages');
+
+class InternalServerError extends Error {
+  constructor(message = messages.internal) {
+    super(message);
+    this.statusCode = 500;
+  }
+}
+
+module.exports = InternalServerError;
