+++ errors/notfounderror.js
@@ -1,0 +1,10 @@
+const { messages } = require('../support/messages');
+
+class NotFoundError extends Error {
+  constructor(message = messages.notFound) {
+    super(message);
+    this.statusCode = 404;
+  }
+}
+
+module.exports = NotFoundError;
