+++ errors/unauthorizederror.js
@@ -1,0 +1,10 @@
+const { messages } = require('../support/messages');
+
+class UnauthorizedError extends Error {
+  constructor(message = messages.unauthorized) {
+    super(message);
+    this.statusCode = 401;
+  }
+}
+
+module.exports = UnauthorizedError;
